+++ app/src/main/java/com/example/weather/data/WeatherData.kt
@@ -7,5 +7,5 @@ data class WeatherData(
     val pressure: Int,
     val humidity: Int,
     val windSpeed: Double,
-    val windDir: Int
+    val windDir: Double
 )
